refactor: thermo-nuclear code-quality pass (decompose macOS hotkey, fix type contract)
Whole-repo maintainability audit (12 finder agents + adversarial verification).
21 candidate findings rejected as thin/speculative/boundary-violating; the 5
high-conviction, behavior-preserving fixes below were applied. The full 353-test
suite stays green and `ty` diagnostics drop 42 -> 37 with no new ones.

desktop/hotkey/macos.py: 702 -> 288 lines. The two monolithic string builders
(a 130-line shell f-string and a 339-line embedded-Swift string) are extracted
to editable resource files (resources/toggle_wrapper.sh, resources/macos_hotkey.swift)
and rendered via simple @@TOKEN@@ substitution. Output is byte-identical for every
case (default, custom timeouts, settings-path, Windows-path, quote-escaping),
verified with a capture/diff oracle and confirmed loadable from an installed wheel.

service/types.py: fix the TranscribeResponse TypedDict contract. The subclass
declared its added keys as required (so conditionally-set `cleanup` errored) and
never declared asr_backend/asr_model/cleanup_backend (which the serializer always
writes); `shell` was typed non-optional though shell_metadata() returns None.
Made the subclass total=False, declared the ASR keys, and widened `shell`. No
runtime change — resolves 5 ty errors in serialize.py.

daemon/windows.py: route log-path resolution through the canonical logs_dir()
helper instead of re-implementing user_log_dir(LOG_DIR_NAME, ...) inline, matching
the Linux and macOS daemons.

daemon/__init__.py: stop re-exporting test-only platform installers
(build_systemd_unit, install_linux_daemon, install_macos_daemon) from the public
API; tests import them from their platform submodules (matching install_windows_daemon).
