+++ transclip/daemon/__init__.py
@@ -2,8 +2,6 @@ from __future__ import annotations
 
 from transclip.product import SERVICE_NAME
 
-from . import linux as linux_daemon
-from . import macos as macos_daemon
 from .common import CommandResult, ServiceState, logs_dir, run_command, service_command, toggle_log_path
 from .lifecycle import (
     install_daemon,
@@ -19,20 +17,17 @@ from .status import (
     stream_logs,
 )
 
-build_systemd_unit = linux_daemon.build_systemd_unit
-install_linux_daemon = linux_daemon.install_linux_daemon
-install_macos_daemon = macos_daemon.install_macos_daemon
-
+# Platform-specific installers (build_systemd_unit, install_linux_daemon,
+# install_macos_daemon, …) are intentionally not re-exported here. Import them
+# from their platform submodules (transclip.daemon.linux/macos/windows) for
+# tests and adapter work; the public surface is the dispatch layer below.
 __all__ = [
     "SERVICE_NAME",
     "CommandResult",
     "ServiceState",
     "append_toggle_log",
-    "build_systemd_unit",
     "collect_status",
     "install_daemon",
-    "install_linux_daemon",
-    "install_macos_daemon",
     "last_toggle_log_event",
     "logs_dir",
     "run_command",
