+++ transclip/desktop/hotkey/macos.py
@@ -7,6 +7,8 @@ import subprocess
 import sys
 from collections.abc import Callable
 from dataclasses import dataclass
+from functools import cache
+from importlib.resources import files
 from pathlib import Path, PurePath
 
 from transclip.daemon.common import CommandResult, logs_dir, repo_root
@@ -85,120 +87,18 @@ def build_macos_toggle_wrapper(
     if settings_path:
         cli += f" --settings {shlex.quote(service_settings_path(settings_path))}"
     restart_command = f'cd {shlex.quote(_macos_path_text(repo_root()))} && {cli} restart >> "$LOG" 2>&1'
-    return f"""#!/bin/sh
-set -u
-
-LOG={shlex.quote(_macos_path_text(log_path))}
-STATE={shlex.quote(_macos_path_text(state_path))}
-BASE={shlex.quote(base_url)}
-MAX_SECONDS={int(stop_timeout_seconds)}
-STALE_LOCK_SECONDS={int(stale_lock_seconds)}
-LOCK=/tmp/transclip-toggle.lock
-
-mkdir -p "$(dirname "$LOG")"
-mkdir -p "$(dirname "$STATE")"
-
-write_state() {{
-  state=$1
-  detail=$2
-  printf '%s\\t%s\\t%s\\n' "$(date '+%Y-%m-%dT%H:%M:%S%z')" "$state" "$detail" > "$STATE"
-  printf '%s\\n' "$(date '+%Y-%m-%dT%H:%M:%S%z') state=${{state}} ${{detail}}" >> "$LOG"
-}}
-
-kill_process_tree() {{
-  pid=$1
-  for child in $(pgrep -P "$pid" 2>/dev/null || true); do
-    kill_process_tree "$child"
-  done
-  kill "$pid" 2>/dev/null || true
-}}
-
-clear_stale_lock_owner() {{
-  owner_pid=$(cat "$LOCK/pid" 2>/dev/null || true)
-  case "$owner_pid" in
-    ''|*[!0-9]*)
-      return
-      ;;
-  esac
-  owner_command=$(ps -p "$owner_pid" -o command= 2>/dev/null || true)
-  case "$owner_command" in
-    *transclip-toggle*)
-      write_state recovering "Clearing stale action"
-      printf '%s\\n' "$(date '+%Y-%m-%dT%H:%M:%S%z') killing stale TransClip action pid=${{owner_pid}}" >> "$LOG"
-      kill_process_tree "$owner_pid"
-      ;;
-  esac
-}}
-
-printf '%s\\n' "$(date '+%Y-%m-%dT%H:%M:%S%z') wrapper invoked" >> "$LOG"
-write_state shortcut "Starting recording"
-
-if ! mkdir "$LOCK" 2>/dev/null; then
-  now=$(date +%s)
-  lock_mtime=$(stat -f %m "$LOCK" 2>/dev/null || printf '0')
-  lock_age=$((now - lock_mtime))
-  if [ "$lock_age" -lt "$STALE_LOCK_SECONDS" ]; then
-    write_state busy "Previous action still running"
-    printf '%s\\n' \
-      "$(date '+%Y-%m-%dT%H:%M:%S%z') ignored: previous TransClip action still running (${{lock_age}}s)" \
-      >> "$LOG"
-    exit 0
-  fi
-  printf '%s\\n' "$(date '+%Y-%m-%dT%H:%M:%S%z') clearing stale TransClip action lock (${{lock_age}}s)" >> "$LOG"
-  clear_stale_lock_owner
-  rm -rf "$LOCK"
-  if ! mkdir "$LOCK" 2>/dev/null; then
-    printf '%s\\n' "$(date '+%Y-%m-%dT%H:%M:%S%z') ignored: could not acquire TransClip action lock" >> "$LOG"
-    exit 0
-  fi
-fi
-printf '%s\\n' "$$" > "$LOCK/pid"
-trap 'rm -rf "$LOCK"' EXIT HUP INT TERM
-
-response=$(curl -sS --max-time 10 -X POST "$BASE/record/start" \
-  -H 'content-type: application/json' --data '{{}}' 2>>"$LOG") || {{
-  write_state error "Start failed"
-  printf '%s\\n' "$(date '+%Y-%m-%dT%H:%M:%S%z') start failed; restarting service" >> "$LOG"
-  {restart_command}
-  exit 0
-}}
-printf '%s\\n' "$response" >> "$LOG"
-case "$response" in
-  *'"already_recording": true'*|*'"already_recording":true'*)
-    already_recording=1
-    ;;
-  *)
-    already_recording=0
-    ;;
-esac
-printf '%s\\n' "$(date '+%Y-%m-%dT%H:%M:%S%z') start already_recording=${{already_recording}}" >> "$LOG"
-
-if [ "$already_recording" = "1" ]; then
-  write_state transcribing "Transcribing"
-  response=$(curl -sS --max-time "$MAX_SECONDS" -X POST "$BASE/record/stop" \
-    -H 'content-type: application/json' --data '{{}}' 2>>"$LOG") || {{
-    write_state error "Stop timed out; restarted"
-    printf '%s\\n' "$(date '+%Y-%m-%dT%H:%M:%S%z') stop failed; restarting service" >> "$LOG"
-    {restart_command}
-    exit 0
-  }}
-  printf '%s\\n' "$response" >> "$LOG"
-  text=$(printf '%s' "$response" | {python} -c \
-    'import json,sys; print(json.load(sys.stdin).get("text", ""), end="")' 2>>"$LOG")
-  if [ -n "$text" ]; then
-    write_state paste_requested "Paste transcript"
-    printf '%s' "$text" | pbcopy
-    printf '%s\\n' "$(date '+%Y-%m-%dT%H:%M:%S%z') copied transcript chars=${{#text}}" >> "$LOG"
-  else
-    write_state finished "No transcript"
-    printf '%s\\n' "$(date '+%Y-%m-%dT%H:%M:%S%z') stop returned no text" >> "$LOG"
-  fi
-else
-  write_state listening "Recording"
-fi
-
-exit 0
-"""
+    return _render_resource(
+        "toggle_wrapper.sh",
+        {
+            "@@LOG@@": shlex.quote(_macos_path_text(log_path)),
+            "@@STATE@@": shlex.quote(_macos_path_text(state_path)),
+            "@@BASE@@": shlex.quote(base_url),
+            "@@MAX_SECONDS@@": str(int(stop_timeout_seconds)),
+            "@@STALE_LOCK_SECONDS@@": str(int(stale_lock_seconds)),
+            "@@RESTART_COMMAND@@": restart_command,
+            "@@PYTHON@@": python,
+        },
+    )
 
 
 def build_macos_hotkey_source(
@@ -206,339 +106,13 @@ def build_macos_hotkey_source(
     log_path: Path,
     state_path: Path,
 ) -> str:
-    source = """import ApplicationServices
-import AppKit
-import Carbon
-import Foundation
-
-let logPath = "@@LOG_PATH@@"
-let wrapperPath = "@@WRAPPER_PATH@@"
-let statePath = "@@STATE_PATH@@"
-let spaceKeyCode: Int64 = 49
-
-func log(_ message: String) {
-    let formatter = ISO8601DateFormatter()
-    let line = "\\(formatter.string(from: Date())) \\(message)\\n"
-    guard let data = line.data(using: .utf8) else { return }
-
-    if FileManager.default.fileExists(atPath: logPath),
-       let handle = try? FileHandle(forWritingTo: URL(fileURLWithPath: logPath)) {
-        defer { try? handle.close() }
-        _ = try? handle.seekToEnd()
-        _ = try? handle.write(contentsOf: data)
-    } else {
-        try? data.write(to: URL(fileURLWithPath: logPath))
-    }
-}
-
-func writeStateFile(_ state: String, _ detail: String) {
-    let formatter = ISO8601DateFormatter()
-    let line = "\\(formatter.string(from: Date()))\\t\\(state)\\t\\(detail)\\n"
-    try? line.write(to: URL(fileURLWithPath: statePath), atomically: true, encoding: .utf8)
-}
-
-func postCommandV() {
-    let source = CGEventSource(stateID: .hidSystemState)
-    let commandKeyCode: CGKeyCode = 55
-    let vKeyCode: CGKeyCode = 9
-
-    let commandDown = CGEvent(keyboardEventSource: source, virtualKey: commandKeyCode, keyDown: true)
-    commandDown?.flags = .maskCommand
-    commandDown?.post(tap: .cghidEventTap)
-    usleep(20_000)
-
-    let vDown = CGEvent(keyboardEventSource: source, virtualKey: vKeyCode, keyDown: true)
-    vDown?.flags = .maskCommand
-    vDown?.post(tap: .cghidEventTap)
-    usleep(20_000)
-
-    let vUp = CGEvent(keyboardEventSource: source, virtualKey: vKeyCode, keyDown: false)
-    vUp?.flags = .maskCommand
-    vUp?.post(tap: .cghidEventTap)
-    usleep(20_000)
-
-    let commandUp = CGEvent(keyboardEventSource: source, virtualKey: commandKeyCode, keyDown: false)
-    commandUp?.post(tap: .cghidEventTap)
-}
-
-class HotkeyStatus: NSObject {
-    let statusItem: NSStatusItem
-    let menu = NSMenu()
-    let statusMenuItem: NSMenuItem
-    var lastStateLine = ""
-    var readyResetTimer: Timer?
-
-    override init() {
-        statusItem = NSStatusBar.system.statusItem(withLength: NSStatusItem.variableLength)
-        statusMenuItem = NSMenuItem(title: "TransClip: Ready", action: nil, keyEquivalent: "")
-        super.init()
-
-        statusMenuItem.isEnabled = false
-        menu.addItem(statusMenuItem)
-        menu.addItem(NSMenuItem.separator())
-
-        let openToggleLog = NSMenuItem(
-            title: "Open toggle log",
-            action: #selector(openToggleLog(_:)),
-            keyEquivalent: ""
-        )
-        openToggleLog.target = self
-        menu.addItem(openToggleLog)
-
-        let openHotkeyLog = NSMenuItem(
-            title: "Open hotkey log",
-            action: #selector(openHotkeyLog(_:)),
-            keyEquivalent: ""
-        )
-        openHotkeyLog.target = self
-        menu.addItem(openHotkeyLog)
-
-        menu.addItem(NSMenuItem.separator())
-        let quit = NSMenuItem(title: "Quit TransClip Hotkey", action: #selector(quit(_:)), keyEquivalent: "q")
-        quit.target = self
-        menu.addItem(quit)
-
-        statusItem.menu = menu
-        setStatus("ready", "Ready")
-    }
-
-    func setStatus(_ state: String, _ detail: String) {
-        if Thread.isMainThread {
-            applyStatus(state, detail)
-        } else {
-            DispatchQueue.main.async {
-                self.applyStatus(state, detail)
-            }
-        }
-    }
-
-    private func applyStatus(_ state: String, _ detail: String) {
-        readyResetTimer?.invalidate()
-        readyResetTimer = nil
-
-        let title: String
-        let fallback: String
-
-        switch state {
-        case "shortcut":
-            title = "TC..."
-            fallback = "Shortcut received"
-        case "busy":
-            title = "TC..."
-            fallback = "Already working"
-        case "recovering":
-            title = "TC..."
-            fallback = "Recovering"
-        case "listening":
-            title = "REC"
-            fallback = "Recording"
-        case "transcribing":
-            title = "TXT..."
-            fallback = "Transcribing"
-        case "pasting":
-            title = "PST..."
-            fallback = "Pasting transcript"
-        case "paste_requested":
-            title = "PST..."
-            fallback = "Paste transcript"
-        case "finished":
-            title = "OK"
-            fallback = "Finished"
-        case "ready":
-            title = "TC"
-            fallback = "Ready"
-        case "error":
-            title = "TC!"
-            fallback = "Error"
-        default:
-            title = "TC"
-            fallback = "Ready"
-        }
-
-        let message = detail.isEmpty ? fallback : detail
-        statusItem.button?.attributedTitle = NSAttributedString(
-            string: title,
-            attributes: [
-                .foregroundColor: color(for: state),
-                .font: NSFont.monospacedSystemFont(ofSize: NSFont.systemFontSize, weight: .semibold),
-            ]
-        )
-        statusItem.button?.toolTip = "TransClip: \\(message)"
-        statusMenuItem.title = "TransClip: \\(message)"
-
-        if state == "finished" {
-            readyResetTimer = Timer.scheduledTimer(withTimeInterval: 2.5, repeats: false) { [weak self] _ in
-                self?.setStatus("ready", "Ready")
-            }
-        }
-    }
-
-    private func color(for state: String) -> NSColor {
-        switch state {
-        case "shortcut", "busy", "recovering":
-            return .systemYellow
-        case "listening":
-            return .systemOrange
-        case "transcribing":
-            return .systemPurple
-        case "pasting", "paste_requested":
-            return .systemTeal
-        case "finished":
-            return .systemGreen
-        case "ready":
-            return .labelColor
-        case "error":
-            return .systemRed
-        default:
-            return .labelColor
-        }
-    }
-
-    @objc func pollState(_ timer: Timer) {
-        guard let line = try? String(contentsOfFile: statePath, encoding: .utf8)
-            .trimmingCharacters(in: .whitespacesAndNewlines),
-              !line.isEmpty,
-              line != lastStateLine else {
-            return
-        }
-
-        lastStateLine = line
-        let parts = line.components(separatedBy: "\\t")
-        let state = parts.count > 1 ? parts[1] : "ready"
-        let detail = parts.count > 2 ? parts[2] : state
-        if state == "paste_requested" {
-            performPasteRequest(detail)
-            return
-        }
-        setStatus(state, detail)
-    }
-
-    func performPasteRequest(_ detail: String) {
-        setStatus("pasting", detail.isEmpty ? "Pasting transcript" : detail)
-        log("paste requested by wrapper")
-        DispatchQueue.main.asyncAfter(deadline: .now() + 0.15) {
-            postCommandV()
-            log("posted Command+V")
-            writeStateFile("finished", "Pasted")
-            self.setStatus("finished", "Pasted")
-        }
-    }
-
-    @objc func openToggleLog(_ sender: Any?) {
-        let toggleLogPath = logPath.replacingOccurrences(
-            of: "hotkey.log",
-            with: "toggle-record.log"
-        )
-        NSWorkspace.shared.open(URL(fileURLWithPath: toggleLogPath))
-    }
-
-    @objc func openHotkeyLog(_ sender: Any?) {
-        NSWorkspace.shared.open(URL(fileURLWithPath: logPath))
-    }
-
-    @objc func quit(_ sender: Any?) {
-        NSApplication.shared.terminate(nil)
-    }
-}
-
-let app = NSApplication.shared
-app.setActivationPolicy(.accessory)
-let hotkeyStatus = HotkeyStatus()
-Timer.scheduledTimer(
-    timeInterval: 0.25,
-    target: hotkeyStatus,
-    selector: #selector(HotkeyStatus.pollState(_:)),
-    userInfo: nil,
-    repeats: true
-)
-
-func runWrapper() {
-    hotkeyStatus.setStatus("shortcut", "Shortcut received")
-    let process = Process()
-    process.executableURL = URL(fileURLWithPath: "/bin/sh")
-    process.arguments = ["-lc", wrapperPath]
-    do {
-        try process.run()
-        log("launched wrapper pid=\\(process.processIdentifier)")
-    } catch {
-        log("failed to launch wrapper: \\(error)")
-    }
-}
-
-let promptKey = kAXTrustedCheckOptionPrompt.takeUnretainedValue() as String
-let trusted = AXIsProcessTrustedWithOptions([promptKey: true] as CFDictionary)
-log("event tap starting axTrusted=\\(trusted)")
-if !trusted {
-    hotkeyStatus.setStatus("error", "Accessibility required")
-}
-
-var activeEventTap: CFMachPort?
-
-func reenableEventTap() {
-    guard let tap = activeEventTap else {
-        log("event tap re-enable skipped; no active tap")
-        return
-    }
-
-    CGEvent.tapEnable(tap: tap, enable: true)
-    log("event tap re-enabled")
-}
-
-let callback: CGEventTapCallBack = { _, type, event, _ in
-    if type == .tapDisabledByTimeout || type == .tapDisabledByUserInput {
-        log("event tap disabled type=\\(type.rawValue)")
-        reenableEventTap()
-        return Unmanaged.passUnretained(event)
-    }
-
-    guard type == .keyDown else {
-        return Unmanaged.passUnretained(event)
-    }
-
-    let keyCode = event.getIntegerValueField(.keyboardEventKeycode)
-    let flags = event.flags
-    let hasOption = flags.contains(.maskAlternate)
-    let hasCommand = flags.contains(.maskCommand)
-    let hasControl = flags.contains(.maskControl)
-
-    if keyCode == spaceKeyCode && hasOption && !hasCommand && !hasControl {
-        log("Option+Space detected")
-        runWrapper()
-        return nil
-    }
-
-    return Unmanaged.passUnretained(event)
-}
-
-let mask = CGEventMask(1 << CGEventType.keyDown.rawValue)
-guard let eventTap = CGEvent.tapCreate(
-    tap: .cgSessionEventTap,
-    place: .headInsertEventTap,
-    options: .defaultTap,
-    eventsOfInterest: mask,
-    callback: callback,
-    userInfo: nil
-) else {
-    log("failed to create event tap; Accessibility/Input Monitoring is required")
-    hotkeyStatus.setStatus("error", "Accessibility required")
-    app.run()
-    exit(0)
-}
-
-activeEventTap = eventTap
-let runLoopSource = CFMachPortCreateRunLoopSource(kCFAllocatorDefault, eventTap, 0)
-CFRunLoopAddSource(CFRunLoopGetCurrent(), runLoopSource, .commonModes)
-CGEvent.tapEnable(tap: eventTap, enable: true)
-log("event tap listening for Option+Space")
-if trusted {
-    hotkeyStatus.setStatus("ready", "Ready")
-}
-app.run()
-"""
-    return (
-        source.replace("@@LOG_PATH@@", _swift_string(_macos_path_text(log_path)))
-        .replace("@@WRAPPER_PATH@@", _swift_string(_macos_path_text(wrapper_path)))
-        .replace("@@STATE_PATH@@", _swift_string(_macos_path_text(state_path)))
+    return _render_resource(
+        "macos_hotkey.swift",
+        {
+            "@@LOG_PATH@@": _swift_string(_macos_path_text(log_path)),
+            "@@WRAPPER_PATH@@": _swift_string(_macos_path_text(wrapper_path)),
+            "@@STATE_PATH@@": _swift_string(_macos_path_text(state_path)),
+        },
     )
 
 
@@ -692,6 +266,18 @@ def _macos_hotkey_executable_path(runtime: PlatformRuntime | None = None) -> Pat
 
 def _macos_hotkey_info_plist_path(runtime: PlatformRuntime | None = None) -> Path:
     return macos_hotkey_app_path(runtime) / "Contents" / "Info.plist"
+
+
+@cache
+def _resource_text(name: str) -> str:
+    return (files(__package__) / "resources" / name).read_text(encoding="utf-8")
+
+
+def _render_resource(name: str, replacements: dict[str, str]) -> str:
+    text = _resource_text(name)
+    for token, value in replacements.items():
+        text = text.replace(token, value)
+    return text
 
 
 def _swift_string(value: str) -> str:
